remove usage of generic error interface
Use the struct directly to represent errors on file & state. The usage of the interface was an issue with gob to save projects state into Bolt project repo: encoding/decoding was not predictable.
Remove the reference of RuleState on RuleStateError: this was unused and caused a infinite cycle with gob encoding (stack overflow panic)