add a tolerance for next backup date
Supposing backup files are upload by a cron task every 24h, checking for a new file exactly after 24h of the previous file was an issue because we would send an obsolete alert whereas a new file is probably being uploaded at the same time. So the tolerance allows for a delay of 2 hours to upload a new file after the latest backup file.